# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TG Restaurant ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lottery/draws:
    get:
      operationId: listLotteryDraws
      tags: [lottery]
      summary: List lottery draws
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Draws
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LotteryDraw"
    post:
      operationId: createLotteryDraw
      tags: [lottery]
      summary: Create lottery draw
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LotteryDrawInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LotteryDraw"
  /lottery/draws/{id}/run:
    post:
      operationId: runLotteryDraw
      tags: [lottery]
      summary: Run lottery draw (pick winners)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Draw results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LotteryResult"
...
components:
  schemas:
...
    LotteryDraw:
      type: object
      required: [id, branchId, drawDate, status]
      properties:
        id:
          type: integer
        branchId:
          type: integer
        drawDate:
          type: string
        eligiblePool:
          type: integer
        status:
          type: string
          enum: [scheduled, completed, cancelled]
        winnersCount:
          type: ["integer", "null"]
    LotteryDrawInput:
      type: object
      required: [branchId, drawDate]
      properties:
        branchId:
          type: integer
        drawDate:
          type: string
        winnersCount:
          type: integer
    LotteryResult:
      type: object
      required: [drawId, winners]
      properties:
        drawId:
          type: integer
        winners:
          type: array
          items:
            $ref: "#/components/schemas/LotteryWinner"
    LotteryWinner:
      type: object
      required: [id, drawId, ticketNumber, prize, notified]
      properties:
        id:
          type: integer
        drawId:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        ticketNumber:
          type: string
        prize:
          type: string
        notified:
          type: boolean

    # DASHBOARD
...